Napravljene sve rute u NewsService-u. Sumnjiva validacija PATCH-a.

+++ NewsService/src/main/java/ba/unsa/etf/nwt/NewsService/repositories/NewsRepository.java
@@ -3,6 +3,6 @@ package ba.unsa.etf.nwt.NewsService.repositories;
 import ba.unsa.etf.nwt.NewsService.model.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsRepository extends JpaRepository<News, Integer> {
+public interface NewsRepository extends JpaRepository<News, Long> {
 
 }
