+++ NewsService/src/main/java/ba/unsa/etf/nwt/NewsService/repositories/NotificationsRepository.java
@@ -3,6 +3,6 @@ package ba.unsa.etf.nwt.NewsService.repositories;
 import ba.unsa.etf.nwt.NewsService.model.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationsRepository extends JpaRepository<Notification, Integer> {
+public interface NotificationsRepository extends JpaRepository<Notification, Long> {
 
 }
